move validation to PreparePubsubWrite
