+++ sdks/java/io/google-cloud-platform/src/main/java/org/apache/beam/sdk/io/gcp/pubsub/PreparePubsubWrite.java
@@ -17,9 +17,9 @@
  */
 package org.apache.beam.sdk.io.gcp.pubsub;
 
-import static org.apache.beam.sdk.io.gcp.pubsub.PubsubIO.validatePubsubMessage;
 import static org.apache.beam.vendor.guava.v26_0_jre.com.google.common.base.Preconditions.checkArgument;
 
+import java.util.Map;
 import javax.naming.SizeLimitExceededException;
 import org.apache.beam.sdk.options.ValueProvider;
 import org.apache.beam.sdk.transforms.DoFn;
@@ -27,11 +27,18 @@ import org.apache.beam.sdk.transforms.PTransform;
 import org.apache.beam.sdk.transforms.ParDo;
 import org.apache.beam.sdk.transforms.SerializableFunction;
 import org.apache.beam.sdk.values.PCollection;
+import org.checkerframework.checker.nullness.qual.Nullable;
 
 public class PreparePubsubWrite<InputT>
     extends PTransform<PCollection<InputT>, PCollection<PubsubMessage>> {
   private SerializableFunction<InputT, PubsubMessage> formatFunction;
   private ValueProvider<PubsubIO.PubsubTopic> topicValueProvider;
+
+  // See https://cloud.google.com/pubsub/quotas#resource_limits.
+  private static final int PUBSUB_MESSAGE_DATA_MAX_LENGTH = 10 << 20;
+  private static final int PUBSUB_MESSAGE_MAX_ATTRIBUTES = 100;
+  private static final int PUBSUB_MESSAGE_ATTRIBUTE_MAX_KEY_LENGTH = 256;
+  private static final int PUBSUB_MESSAGE_ATTRIBUTE_MAX_VALUE_LENGTH = 1024;
 
   public PreparePubsubWrite(
       ValueProvider<PubsubIO.PubsubTopic> topicValueProvider,
@@ -76,6 +83,48 @@ public class PreparePubsubWrite<InputT>
         throw new IllegalArgumentException(e);
       }
       context.output(outputValue);
+    }
+  }
+
+  public static void validatePubsubMessage(PubsubMessage message)
+      throws SizeLimitExceededException {
+    if (message.getPayload().length > PUBSUB_MESSAGE_DATA_MAX_LENGTH) {
+      throw new SizeLimitExceededException(
+          "Pubsub message data field of length "
+              + message.getPayload().length
+              + " exceeds maximum of "
+              + PUBSUB_MESSAGE_DATA_MAX_LENGTH
+              + ". See https://cloud.google.com/pubsub/quotas#resource_limits");
+    }
+    @Nullable Map<String, String> attributes = message.getAttributeMap();
+    if (attributes != null) {
+      if (attributes.size() > PUBSUB_MESSAGE_MAX_ATTRIBUTES) {
+        throw new SizeLimitExceededException(
+            "Pubsub message contains "
+                + attributes.size()
+                + " attributes which exceeds the maximum of "
+                + PUBSUB_MESSAGE_MAX_ATTRIBUTES
+                + ". See https://cloud.google.com/pubsub/quotas#resource_limits");
+      }
+      for (Map.Entry<String, String> attribute : attributes.entrySet()) {
+        if (attribute.getKey().length() > PUBSUB_MESSAGE_ATTRIBUTE_MAX_KEY_LENGTH) {
+          throw new SizeLimitExceededException(
+              "Pubsub message attribute key "
+                  + attribute.getKey()
+                  + " exceeds the maximum of "
+                  + PUBSUB_MESSAGE_ATTRIBUTE_MAX_KEY_LENGTH
+                  + ". See https://cloud.google.com/pubsub/quotas#resource_limits");
+        }
+        String value = attribute.getValue();
+        if (value.length() > PUBSUB_MESSAGE_ATTRIBUTE_MAX_VALUE_LENGTH) {
+          throw new SizeLimitExceededException(
+              "Pubsub message attribute value starting with "
+                  + value.substring(0, Math.min(256, value.length()))
+                  + " exceeds the maximum of "
+                  + PUBSUB_MESSAGE_ATTRIBUTE_MAX_VALUE_LENGTH
+                  + ". See https://cloud.google.com/pubsub/quotas#resource_limits");
+        }
+      }
     }
   }
 }
