+++ sdks/java/io/google-cloud-platform/src/test/java/org/apache/beam/sdk/io/gcp/pubsub/PubSubWritePayloadTranslationTest.java
@@ -17,9 +17,29 @@
  */
 package org.apache.beam.sdk.io.gcp.pubsub;
 
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
+
+import org.apache.beam.model.pipeline.v1.RunnerApi;
+import org.apache.beam.model.pipeline.v1.RunnerApi.PubSubWritePayload;
+import org.apache.beam.runners.core.construction.Environments;
+import org.apache.beam.runners.core.construction.PTransformTranslation;
+import org.apache.beam.runners.core.construction.SdkComponents;
 import org.apache.beam.sdk.io.gcp.pubsub.PubSubPayloadTranslation.PubSubWritePayloadTranslator;
 import org.apache.beam.sdk.io.gcp.pubsub.PubsubClient.TopicPath;
+import org.apache.beam.sdk.io.gcp.pubsub.PubsubUnboundedSink.PubsubSink;
+import org.apache.beam.sdk.options.ValueProvider;
+import org.apache.beam.sdk.options.ValueProvider.NestedValueProvider;
+import org.apache.beam.sdk.options.ValueProvider.StaticValueProvider;
+import org.apache.beam.sdk.runners.AppliedPTransform;
 import org.apache.beam.sdk.testing.TestPipeline;
+import org.apache.beam.sdk.transforms.Create;
+import org.apache.beam.sdk.transforms.resourcehints.ResourceHints;
+import org.apache.beam.sdk.values.KV;
+import org.apache.beam.sdk.values.PCollection;
+import org.apache.beam.sdk.values.PDone;
+import org.apache.beam.sdk.values.PValues;
+import org.joda.time.Duration;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -39,7 +59,7 @@ public class PubSubWritePayloadTranslationTest {
 
   @Test
   public void testTranslateSinkWithTopic() throws Exception {
-    /* PubsubUnboundedSink pubsubUnboundedSink =
+     PubsubUnboundedSink pubsubUnboundedSink =
         new PubsubUnboundedSink(
             null,
             StaticValueProvider.of(TOPIC),
@@ -51,9 +71,9 @@ public class PubSubWritePayloadTranslationTest {
             Duration.ZERO,
             null);
     PubsubUnboundedSink.PubsubSink pubsubSink = new PubsubSink(pubsubUnboundedSink);
-    PCollection<KV<PubsubIO.PubsubTopic, byte[]>> input =
+    PCollection<KV<String, byte[]>> input =
         pipeline.apply(
-            Create.of(KV.of(PubsubIO.PubsubTopic.fromPath(TOPIC.getPath()), new byte[0])));
+            Create.of(KV.of(TOPIC.getPath(), new byte[0])));
     PDone output = input.apply(pubsubSink);
     AppliedPTransform<?, ?, PubsubSink> appliedPTransform =
         AppliedPTransform.of(
@@ -72,19 +92,19 @@ public class PubSubWritePayloadTranslationTest {
     assertEquals(TOPIC.getFullPath(), payload.getTopic());
     assertTrue(payload.getTopicRuntimeOverridden().isEmpty());
     assertEquals(TIMESTAMP_ATTRIBUTE, payload.getTimestampAttribute());
-    assertEquals(ID_ATTRIBUTE, payload.getIdAttribute());*/
+    assertEquals(ID_ATTRIBUTE, payload.getIdAttribute());
   }
 
   @Test
   public void testTranslateSinkWithTopicOverridden() throws Exception {
-    /* ValueProvider<TopicPath> runtimeProvider = pipeline.newProvider(TOPIC);
+     ValueProvider<TopicPath> runtimeProvider = pipeline.newProvider(TOPIC);
     PubsubUnboundedSink pubsubUnboundedSinkSink =
         new PubsubUnboundedSink(
             null, runtimeProvider, TIMESTAMP_ATTRIBUTE, ID_ATTRIBUTE, 0, 0, 0, Duration.ZERO, null);
     PubsubSink pubsubSink = new PubsubSink(pubsubUnboundedSinkSink);
-    PCollection<KV<PubsubIO.PubsubTopic, byte[]>> input =
+    PCollection<KV<String, byte[]>> input =
         pipeline.apply(
-            Create.of(KV.of(PubsubIO.PubsubTopic.fromPath(TOPIC.getPath()), new byte[0])));
+            Create.of(KV.of(TOPIC.getPath(), new byte[0])));
     PDone output = input.apply(pubsubSink);
     AppliedPTransform<?, ?, PubsubSink> appliedPTransform =
         AppliedPTransform.of(
@@ -105,6 +125,6 @@ public class PubSubWritePayloadTranslationTest {
         payload.getTopicRuntimeOverridden());
     assertTrue(payload.getTopic().isEmpty());
     assertEquals(TIMESTAMP_ATTRIBUTE, payload.getTimestampAttribute());
-    assertEquals(ID_ATTRIBUTE, payload.getIdAttribute());*/
+    assertEquals(ID_ATTRIBUTE, payload.getIdAttribute());
   }
 }
